1.讀書會User端   共用分頁:所有讀書會/我發起的/我參加的   發起人:查看報名情況、報名明細、修改報名狀態   參與者:報名參加、取消報名

+++ src/main/java/bookstore/bean/ClubRegistrationsBean.java
@@ -31,7 +31,7 @@ public class ClubRegistrationsBean {
     private UserBean user; // 關聯報名者
 
     @Column(name = "status", nullable = false)
-    private Short status;
+    private Integer status;
 
     @Column(name = "check_in")
     private Boolean checkIn; // BIT 對應 Boolean
